# api/openapi.yaml
openapi: 3.0.3
info:
  title: Office Tracker API
  description: |
    REST API for tracking office attendance using RFID cards.

    ## Authentication
    If API key authentication is enabled (via `API_KEY` environment variable),
    all endpoints except `/health` require authentication using one of these methods:
    - Header: `X-API-Key: your-api-key`
    - Header: `Authorization: Bearer your-api-key`

    ## CORS
    CORS can be enabled by setting the `CORS_ORIGINS` environment variable.
    - Set to `*` to allow all origins
    - Set to comma-separated list for specific origins (e.g., `http://localhost:3000,https://example.com`)
  version: 1.0.0
  contact:
    name: Office Tracker Team
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/sessions/checkout:
    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Check out a user
      description: Close the active session for a user
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSessionActionRequest"
      responses:
        "200":
          description: User checked out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSessionActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/sessions:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: List sessions with filters
      description: Retrieve sessions with optional filtering by date range, user name, sorting, and pagination
      tags:
        - Sessions
      parameters:
        - name: from
          in: query
          required: false
          description: Filter sessions from this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Filter sessions until this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-28T23:59:59Z"
        - name: name
          in: query
          required: false
          description: Filter sessions by user name (partial match)
          schema:
            type: string
            example: John
        - name: user_id
          in: query
          required: false
          description: Filter sessions by user ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: discord_id
          in: query
          required: false
          description: Filter sessions by Discord ID
          schema:
            type: string
            example: "123456789012345678"
        - name: status
          in: query
          required: false
          description: Session status filter
          schema:
            type: string
            enum: [all, active, completed]
            example: completed
        - name: active_only
          in: query
          required: false
          description: Legacy status alias (`true` = active, `false` = completed)
          schema:
            type: boolean
            example: false
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 100)
          schema:
            type: integer
            minimum: 1
            example: 100
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order (default 'desc')
          schema:
            type: string
            enum: [asc, desc]
            example: desc
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (default 'check_in')
          schema:
            type: string
            enum: [check_in, check_out, user_name]
            example: check_in
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Delete sessions (bulk)
      description: |
        Delete sessions matching filters.
        At least one filter must be provided.
        When `limit` or `offset` are supplied, only that ordered slice is deleted.
      tags:
        - Sessions
      parameters:
        - name: from
          in: query
          required: false
          description: Delete sessions from this datetime (RFC3339)
          schema:
            type: string
            format: date-time
            example: "2026-02-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Delete sessions until this datetime (RFC3339)
          schema:
            type: string
            format: date-time
            example: "2026-02-28T23:59:59Z"
        - name: name
          in: query
          required: false
          description: Delete sessions by user name (partial match)
          schema:
            type: string
            example: John
        - name: user_id
          in: query
          required: false
          description: Delete sessions for a specific user ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: discord_id
          in: query
          required: false
          description: Delete sessions for a specific Discord ID
          schema:
            type: string
            example: "123456789012345678"
        - name: status
          in: query
          required: false
          description: Delete only sessions with the selected status
          schema:
            type: string
            enum: [all, active, completed]
            example: completed
        - name: active_only
          in: query
          required: false
          description: Legacy status alias (`true` = active, `false` = completed)
          schema:
            type: boolean
            example: false
        - name: limit
          in: query
          required: false
          description: Maximum number of matching rows to delete from the ordered result set
          schema:
            type: integer
            minimum: 1
            example: 100
        - name: offset
          in: query
          required: false
          description: Number of matching rows to skip before deleting
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order applied before pagination (default `desc`)
          schema:
            type: string
            enum: [asc, desc]
            example: desc
        - name: sort_by
          in: query
          required: false
          description: Field used for ordered bulk delete pagination
          schema:
            type: string
            enum: [check_in, check_out, user_name]
            example: check_in
      responses:
        "200":
          description: Sessions deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    format: int64
                    example: 15
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional, configured via API_KEY env var)
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (alternative to X-API-Key)
  schemas:
...
    SessionResponse:
      type: object
      description: Session response DTO with user information
      properties:
        id:
          type: integer
          format: int64
          description: Session ID
          example: 42
        user_id:
          type: integer
          format: int64
          description: User ID
          example: 7
        user_name:
          type: string
          description: Name of the user
          example: John Doe
        check_in:
          type: string
          format: date-time
          description: Check-in timestamp
          example: "2026-02-15T09:00:00Z"
        check_out:
          type: string
          format: date-time
          nullable: true
          description: Check-out timestamp (null if still active)
          example: "2026-02-15T17:00:00Z"
        active:
          type: boolean
          description: Whether the session is currently active
          example: false
      required:
        - id
        - user_id
        - user_name
        - check_in
        - active
...
    UserSessionActionRequest:
      type: object
      description: Request body for checking in/out a user
      properties:
        user_id:
          type: integer
          format: int64
          example: 1
      required:
        - user_id
    UserSessionActionResponse:
      type: object
      properties:
        status:
          type: string
          example: checked-in
      required:
        - status
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        text/plain:
          schema:
            type: string
            example: invalid request
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: database error